pyproject.toml
[tool.poetry]
name = "gitfix"
version = "0.1.0"
description = ""
authors = ["Devansh Singh <devanshamity@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.27.0"
bs4 = "^0.0.2"
lxml = "^5.2.2"
chromadb = "^0.5.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
